refactor(hooks): extract HTTP hook execution from EE into CE
